Lesson 5

+++ app/controllers/questions_controller.rb
@@ -1,12 +1,16 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
   expose :questions, -> { Question.all }
   expose :question
+  expose :answer, -> { question.answers.new }
 
   def create
+    @exposed_question = current_user.questions.new(question_params)
+
     if question.save
-      redirect_to question
+      redirect_to question, notice: "Your question successfully created."
     else
       render :new
     end
@@ -21,8 +25,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to question, notice: "Question was successfully deleted"
+    else
+      redirect_to question, notice: "Can't delete the question"
+    end
   end
 
   private
